// src/app/shared/models/user.model.ts
export class User {
    userId: number;
    email: string;
    password: string;
    firstName: string;
    lastName: string;
    nativeUser: any[];
    emailAddress: string;
    lastLoginDt: string;
    endDt: string;
    enabled: string;
}
export class FieldMapping {
    id: number;
    jobTitle: string;
    jobCode: string;
    payPlan: string;
    grade: string;
    fpl: string;
    occSeries: string;
}